anomaly_detection/anomaly_detection.py
import os
import boto3
import pandas as pd
from sklearn.ensemble import IsolationForest

# Use environment variables for configuration
S3_BUCKET = os.getenv('S3_BUCKET')
S3_PREFIX = os.getenv('S3_PREFIX')

def load_data_from_s3():
    s3 = boto3.client('s3')
    # Implementation to load data from S3
    # This is a placeholder and needs to be implemented based on your S3 structure
    pass

def train_model(data):
    model = IsolationForest(contamination=0.1, random_state=42)
    model.fit(data)
    return model

def detect_anomalies(model, data):
    predictions = model.predict(data)
    return [i for i, pred in enumerate(predictions) if pred == -1]

def main():
    # Load data from S3
    data = load_data_from_s3()
    
    # Train model
    model = train_model(data)
    
    # Detect anomalies
    anomalies = detect_anomalies(model, data)
    
    # Process anomalies (e.g., send alerts, log to database, etc.)
    print(f"Detected {len(anomalies)} anomalies")

if __name__ == "__main__":
    main()